requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.49
Flask-Login==0.6.2
python-dotenv==0.19.0
gunicorn==20.1.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pandas==1.5.3
numpy==1.24.3
openpyxl==3.1.2